refact!: Move field methods to `Content\Field`
Resolves #7078
BREAKING CHANGE: Removed `Kirby\Content\Field::$aliases`.
BREAKING CHANGE: Removed `Kirby\Cms\Core::fieldMethods()` and `Kirby\Cms\Core::fieldMethodsAliases()`.
