+++ tests/Cms/CoreTest.php
@@ -103,76 +103,6 @@ class CoreTest extends TestCase
 		$this->assertArrayHasKey('url', $components);
 	}
 
-	public function testFieldMethodAliases(): void
-	{
-		$aliases = $this->core->fieldMethodAliases();
-
-		$this->assertSame('toBool', $aliases['bool']);
-		$this->assertSame('escape', $aliases['esc']);
-		$this->assertSame('toExcerpt', $aliases['excerpt']);
-		$this->assertSame('toFloat', $aliases['float']);
-		$this->assertSame('html', $aliases['h']);
-		$this->assertSame('toInt', $aliases['int']);
-		$this->assertSame('kirbytext', $aliases['kt']);
-		$this->assertSame('kirbytextinline', $aliases['kti']);
-		$this->assertSame('toLink', $aliases['link']);
-		$this->assertSame('markdown', $aliases['md']);
-		$this->assertSame('smartypants', $aliases['sp']);
-		$this->assertSame('isValid', $aliases['v']);
-		$this->assertSame('xml', $aliases['x']);
-	}
-
-	public function testFieldMethods(): void
-	{
-		$methods = $this->core->fieldMethods();
-
-		$this->assertArrayHasKey('isFalse', $methods);
-		$this->assertArrayHasKey('isTrue', $methods);
-		$this->assertArrayHasKey('isValid', $methods);
-		$this->assertArrayHasKey('toBlocks', $methods);
-		$this->assertArrayHasKey('toBool', $methods);
-		$this->assertArrayHasKey('toData', $methods);
-		$this->assertArrayHasKey('toDate', $methods);
-		$this->assertArrayHasKey('toFile', $methods);
-		$this->assertArrayHasKey('toFiles', $methods);
-		$this->assertArrayHasKey('toFloat', $methods);
-		$this->assertArrayHasKey('toInt', $methods);
-		$this->assertArrayHasKey('toLayouts', $methods);
-		$this->assertArrayHasKey('toLink', $methods);
-		$this->assertArrayHasKey('toPage', $methods);
-		$this->assertArrayHasKey('toPages', $methods);
-		$this->assertArrayHasKey('toPages', $methods);
-		$this->assertArrayHasKey('toStructure', $methods);
-		$this->assertArrayHasKey('toTimestamp', $methods);
-		$this->assertArrayHasKey('toUrl', $methods);
-		$this->assertArrayHasKey('toUser', $methods);
-		$this->assertArrayHasKey('toUsers', $methods);
-		$this->assertArrayHasKey('length', $methods);
-		$this->assertArrayHasKey('words', $methods);
-		$this->assertArrayHasKey('words', $methods);
-		$this->assertArrayHasKey('callback', $methods);
-		$this->assertArrayHasKey('escape', $methods);
-		$this->assertArrayHasKey('excerpt', $methods);
-		$this->assertArrayHasKey('html', $methods);
-		$this->assertArrayHasKey('inline', $methods);
-		$this->assertArrayHasKey('kirbytext', $methods);
-		$this->assertArrayHasKey('kirbytextinline', $methods);
-		$this->assertArrayHasKey('kirbytags', $methods);
-		$this->assertArrayHasKey('lower', $methods);
-		$this->assertArrayHasKey('markdown', $methods);
-		$this->assertArrayHasKey('nl2br', $methods);
-		$this->assertArrayHasKey('query', $methods);
-		$this->assertArrayHasKey('replace', $methods);
-		$this->assertArrayHasKey('short', $methods);
-		$this->assertArrayHasKey('slug', $methods);
-		$this->assertArrayHasKey('smartypants', $methods);
-		$this->assertArrayHasKey('split', $methods);
-		$this->assertArrayHasKey('upper', $methods);
-		$this->assertArrayHasKey('widont', $methods);
-		$this->assertArrayHasKey('xml', $methods);
-		$this->assertArrayHasKey('yaml', $methods);
-	}
-
 	public function testFieldMixins(): void
 	{
 		$mixins = $this->core->fieldMixins();
